+++ src/Products/PloneMeeting/vocabularies.py
@@ -2525,7 +2525,6 @@ SelectableCommitteesVocabularyFactory = SelectableCommitteesVocabulary()
 
 
 class SelectableCommitteesAcronymsVocabulary(SelectableCommitteesVocabulary):
-    implements(IVocabularyFactory)
 
     def __call__(self, context, term_title_attr="acronym"):
         """ """
@@ -2537,7 +2536,6 @@ SelectableCommitteesAcronymsVocabularyFactory = SelectableCommitteesAcronymsVoca
 
 
 class ItemSelectableCommitteesVocabulary(SelectableCommitteesVocabulary):
-    implements(IVocabularyFactory)
 
     def _get_stored_values(self):
         """ """
@@ -2562,7 +2560,6 @@ ItemSelectableCommitteesVocabularyFactory = ItemSelectableCommitteesVocabulary()
 
 
 class MeetingSelectableCommitteesVocabulary(SelectableCommitteesVocabulary):
-    implements(IVocabularyFactory)
 
     def _get_stored_values(self):
         """ """
@@ -2582,6 +2579,74 @@ class MeetingSelectableCommitteesVocabulary(SelectableCommitteesVocabulary):
 
 
 MeetingSelectableCommitteesVocabularyFactory = MeetingSelectableCommitteesVocabulary()
+
+
+class OtherMCsClonableToVocabulary(object):
+    """Vocabulary listing other MeetingConfigs clonable to."""
+
+    implements(IVocabularyFactory)
+
+    def __call___cachekey(method, self, context, term_title=None):
+        '''cachekey method for self.__call__.'''
+        tool = api.portal.get_tool('portal_plonemeeting')
+        cfg = tool.getMeetingConfig(context)
+        return repr(cfg), term_title
+
+    @ram.cache(__call___cachekey)
+    def OtherMCsClonableToVocabulary__call__(self, context, term_title=None):
+        """ """
+        tool = api.portal.get_tool('portal_plonemeeting')
+        cfg = tool.getMeetingConfig(context)
+        terms = []
+        import logging
+        logging.getLogger('gna').info('OtherMCsClonableToVocabulary')
+        # XXX not a real usecase, disabled value on an item?  Would complicate caching
+        # cfg_ids = [mc['meeting_config'] for mc in cfg.getMeetingConfigsToCloneTo()]
+        # if context.__class__.__name__ == 'MeetingItem':
+        #     cfg_ids = list(set(cfg_ids).union(context.getOtherMeetingConfigsClonableTo()))
+        for mctct in cfg.getMeetingConfigsToCloneTo():
+            terms.append(SimpleTerm(mctct['meeting_config'],
+                                    mctct['meeting_config'],
+                                    term_title or getattr(tool, mctct['meeting_config']).Title()))
+        return SimpleVocabulary(terms)
+
+    # do ram.cache have a different key name
+    __call__ = OtherMCsClonableToVocabulary__call__
+
+
+OtherMCsClonableToVocabularyFactory = OtherMCsClonableToVocabulary()
+
+
+class OtherMCsClonableToEmergencyVocabulary(OtherMCsClonableToVocabulary):
+    """Vocabulary listing other MeetingConfigs clonable to emergency."""
+
+    implements(IVocabularyFactory)
+
+    def __call__(self, context, term_title=None):
+        """ """
+        term_title = translate('Emergency while presenting in other MC',
+                               domain='PloneMeeting',
+                               context=context.REQUEST)
+        return super(OtherMCsClonableToEmergencyVocabulary, self).__call__(context, term_title)
+
+
+OtherMCsClonableToEmergencyVocabularyFactory = OtherMCsClonableToEmergencyVocabulary()
+
+
+class OtherMCsClonableToPrivacyVocabulary(OtherMCsClonableToVocabulary):
+    """Vocabulary listing other MeetingConfigs clonable to privacy."""
+
+    implements(IVocabularyFactory)
+
+    def __call__(self, context, term_title=None):
+        """ """
+        term_title = translate('Secret while presenting in other MC?',
+                               domain='PloneMeeting',
+                               context=context.REQUEST)
+        return super(OtherMCsClonableToPrivacyVocabulary, self).__call__(context, term_title)
+
+
+OtherMCsClonableToPrivacyVocabularyFactory = OtherMCsClonableToPrivacyVocabulary()
 
 
 class ContainedAnnexesVocabulary(object):